+++ src/parser/enrichers/monster/BladeOfDisaster/_module.mjs
@@ -1,0 +1,2 @@
+export { default as MoveAndAttack } from "./MoveAndAttack.mjs";
+