+++ src/parser/enrichers/monster/ConjuredConstruct/OrderlyWard.mjs
@@ -1,0 +1,23 @@
+/* eslint-disable class-methods-use-this */
+// import { utils } from "../../../../lib/_module.mjs";
+import DDBEnricherData from "../../data/DDBEnricherData.mjs";
+
+export default class OrderlyWard extends DDBEnricherData {
+  get type() {
+    return "heal";
+  }
+
+  get activity() {
+    return {
+      data: {
+        healing: DDBEnricherData.basicDamagePart({
+          number: 1,
+          denomination: 6,
+          bonus: "@flags.dnd5e.summon.mod",
+        }),
+      },
+    };
+  }
+
+
+}