+++ src/parser/enrichers/monster/ConjuredConstruct/_module.mjs
@@ -1,0 +1,2 @@
+export { default as OrderlyWard } from "./OrderlyWard.mjs";
+